Fixing back animation
