+++ lib/login/ui/forget.dart
@@ -4,6 +4,7 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:heroicons/heroicons.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:myecl/auth/providers/openid_provider.dart';
+import 'package:myecl/login/providers/animation_provider.dart';
 import 'package:myecl/login/providers/sign_up_provider.dart';
 import 'package:myecl/login/router.dart';
 import 'package:myecl/login/tools/constants.dart';
@@ -20,6 +21,7 @@ class ForgetPassword extends HookConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final signUpNotifier = ref.watch(signUpProvider.notifier);
     final email = useTextEditingController();
+    final controller = ref.watch(backgroundAnimationProvider);
     void displayToastWithContext(TypeMsg type, String msg) {
       displayToast(context, type, msg);
     }
@@ -42,6 +44,7 @@ class ForgetPassword extends HookConsumerWidget {
                 child: GestureDetector(
                   onTap: () {
                     QR.to(LoginRouter.root);
+                    controller?.reverse();
                   },
                   child: const HeroIcon(
                     HeroIcons.chevronLeft,
@@ -93,7 +96,7 @@ class ForgetPassword extends HookConsumerWidget {
                     const SizedBox(
                       height: 30,
                     ),
-                    SignUpBar(
+                    SignInUpBar(
                       label: LoginTextConstants.recover,
                       isLoading: ref.watch(loadingrovider).when(
                           data: (data) => data,
@@ -125,6 +128,7 @@ class ForgetPassword extends HookConsumerWidget {
                               splashColor: const Color.fromRGBO(255, 255, 255, 1),
                               onTap: () {
                                 QR.to(LoginRouter.root);
+                                controller?.reverse();
                               },
                               child: const Text(
                                 LoginTextConstants.signIn,
